simplify, rename, and comment supplier used to detect recursive ACLs
