+++ projects/batfish-common-protocol/src/test/java/org/batfish/common/util/NonRecursiveSupplierTest.java
@@ -4,20 +4,20 @@ import com.google.common.base.Suppliers;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
-import org.batfish.common.util.NonReentrantSupplier.NonReentrantSupplierException;
+import org.batfish.common.util.NonRecursiveSupplier.NonRecursiveSupplierException;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
-public class NonReentrantSupplierTest {
+public class NonRecursiveSupplierTest {
   @Rule public ExpectedException exception = ExpectedException.none();
 
   @Test
   public void testNonReentrantSupplier() {
     Map<String, Supplier<String>> suppliers = new HashMap<>();
-    suppliers.put("foo", new NonReentrantSupplier<>(() -> suppliers.get("bar").get()));
-    suppliers.put("bar", new NonReentrantSupplier<>(() -> suppliers.get("foo").get()));
-    exception.expect(NonReentrantSupplierException.class);
+    suppliers.put("foo", new NonRecursiveSupplier<>(() -> suppliers.get("bar").get()));
+    suppliers.put("bar", new NonRecursiveSupplier<>(() -> suppliers.get("foo").get()));
+    exception.expect(NonRecursiveSupplierException.class);
     suppliers.get("foo").get();
   }
 
@@ -25,10 +25,10 @@ public class NonReentrantSupplierTest {
   public void testMemoize() {
     Map<String, Supplier<String>> suppliers = new HashMap<>();
     suppliers.put(
-        "foo", Suppliers.memoize(new NonReentrantSupplier<>(() -> suppliers.get("bar").get())));
+        "foo", Suppliers.memoize(new NonRecursiveSupplier<>(() -> suppliers.get("bar").get())));
     suppliers.put(
-        "bar", Suppliers.memoize(new NonReentrantSupplier<>(() -> suppliers.get("foo").get())));
-    exception.expect(NonReentrantSupplierException.class);
+        "bar", Suppliers.memoize(new NonRecursiveSupplier<>(() -> suppliers.get("foo").get())));
+    exception.expect(NonRecursiveSupplierException.class);
     suppliers.get("foo").get();
   }
 }
